refactor: move code to webui.py
